Assignment of Element Series and Color Coding
Not the most elegant solution, but there is a variety of ways the periodic table is organized making this a rather difficult project for my portfolio. However, this is more about development and expanding my skillsets, so there naturally will be some flaws with this table due to my lack of conventional education in Chemistry.